Integrate License Manager plugin and other dev changes

+++ wp-express-checkout/admin/includes/class-products-list.php
@@ -61,7 +61,7 @@ class Products_List {
 					),
 					array(
 						'name'            => get_the_title( $post_id ),
-						'price'           => get_post_meta( $post_id, 'ppec_product_price', true ),
+						'price'           => (float) get_post_meta( $post_id, 'ppec_product_price', true ),
 						'shipping'        => ( '' === get_post_meta( $post_id, 'wpec_product_shipping', true ) ) ? $main->get_setting( 'shipping' ) : get_post_meta( $post_id, 'wpec_product_shipping', true ),
 						'tax'             => ( '' === get_post_meta( $post_id, 'wpec_product_tax', true ) ) ? $main->get_setting( 'tax' ) : get_post_meta( $post_id, 'wpec_product_tax', true ),
 						'quantity'        => get_post_meta( $post_id, 'ppec_product_quantity', true ),
